Ready: 151 refactor classifier (#208)
* Refactors ClassificationReport and ConfusionMatrix.

Refactors tests for above
Updates ipynb example in nealhumphrey

Includes:
- partial bug-fix for 3+ class data used w/ the classification report (need to get shape of image right, but no longer errors out)
- enhances the classificaiton report to use dynamic text color

TODO:
Need to fix examples.ipynb to import properly (can't do w/out internet)

* - Moves over the rocauc and class_balance

- adds test for class_balance

TODO:
- Update the examples.ipynb (and any other examples) to draw from the right place
- Documentation comment as noted by Ben

Extra:
- Allow the class balance to avoid using a model (just data)

* Renaming of the outdated classifier module. This is destined to be deleted once everything has been fully tested/integrated.

* Tries adding these to commit

* Corrects the namespace import issue

* Update autodoc, remove old code.

* Updated docstrings
