+++ docs/api/yellowbrick.classifier.rst
@@ -1,0 +1,45 @@
+yellowbrick.classifier package
+=============================
+
+Submodules
+----------
+
+yellowbrick.classifier.base module
+---------------------------------
+
+.. automodule:: yellowbrick.classifier.base
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+yellowbrick.classifier.class_balance module
+--------------------------------------
+
+.. automodule:: yellowbrick.classifier.class_balance
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+yellowbrick.classifier.classification_report module
+-----------------------------------
+
+.. automodule:: yellowbrick.classifier.classification_report
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+yellowbrick.classifier.confusion_matrix module
+-----------------------------------
+
+.. automodule:: yellowbrick.classifier.confusion_matrix
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+yellowbrick.classifier.rocauc module
+-----------------------------------
+
+.. automodule:: yellowbrick.classifier.rocauc
+    :members:
+    :undoc-members:
+    :show-inheritance:
