+++ yellowbrick/classifier/__init__.py
@@ -1,0 +1,15 @@
+# yellowbrick.classifier
+# Visualizers for Classification analysis and diagnostics
+
+##########################################################################
+## Imports
+##########################################################################
+
+## Hoist visualizers into the classifier namespace
+from ..base import ScoreVisualizer
+from .base import ClassificationScoreVisualizer
+from .class_balance import ClassBalance
+from .classification_report import ClassificationReport, classification_report
+from .confusion_matrix import ConfusionMatrix
+from .rocauc import ROCAUC, roc_auc
+
